Check if internet is reachable
I added a simple decorator as well as checking for internet connection with ping to see if whether the host is down or internet is not reachable. Optionally, retry keyword argument can be used to retry to reach internet in case it fails and opt_exception to raise an optional exception.